# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synapse Enterprise Intelligence API",
    "description": "Graphiti-backed temporal knowledge graph platform. Ingests enterprise documents, extracts a closed-vocabulary ontology, resolves entities, and maintains bi-temporal fact validity.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/snapshot/generate": {
      "post": {
        "summary": "Generate Snapshot",
        "operationId": "generate_snapshot_api_v1_snapshot_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Generate Snapshot Api V1 Snapshot Generate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/snapshots/{version}/download": {
      "get": {
        "summary": "Download Snapshot",
        "operationId": "download_snapshot_api_v1_snapshots__version__download_get",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/snapshots/{version}/notify": {
      "post": {
        "summary": "Notify Downstream",
        "operationId": "notify_downstream_api_v1_snapshots__version__notify_post",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Notify Downstream Api V1 Snapshots  Version  Notify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SnapshotRequest": {
        "properties": {
          "tenants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Tenants"
          },
          "factScope": {
            "type": "string",
            "title": "Factscope",
            "default": "current"
          },
          "reportScope": {
            "type": "string",
            "title": "Reportscope",
            "default": "flagged"
          },
          "accessControlHandling": {
            "type": "string",
            "title": "Accesscontrolhandling"
          },
          "confirmationPhrase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmationphrase"
          }
        },
        "type": "object",
        "required": [
          "tenants",
          "accessControlHandling"
        ],
        "title": "SnapshotRequest",
        "description": "Mirrors the Zod schema in ExportScreen."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}